+++ packages/visual-editor/src/components/contentBlocks/image/Image.tsx
@@ -7,6 +7,7 @@ import {
 import { ComplexImageType, ImageType } from "@yext/pages-components";
 import {
   AssetImageType,
+  TranslatableAssetImage,
   EntityField,
   Image,
   ImgSizesByBreakpoint,
@@ -33,7 +34,9 @@ const LINK_REGEX_VALIDATION = /^(https?:\/\/[^\s]+|\/[^\s]*|#[^\s]*)$/;
 export interface ImageWrapperProps {
   data: {
     /** The image to display. */
-    image: YextEntityField<ImageType | ComplexImageType | AssetImageType>;
+    image: YextEntityField<
+      ImageType | ComplexImageType | TranslatableAssetImage
+    >;
     link?: TranslatableString;
   };
 
@@ -43,7 +46,7 @@ export interface ImageWrapperProps {
   /** @internal Controlled data from the parent section. */
   parentData?: {
     field: string;
-    image: ImageType | ComplexImageType | AssetImageType | undefined;
+    image: ImageType | ComplexImageType | TranslatableAssetImage | undefined;
   };
 
   /** Additional CSS classes to apply to the image. */
@@ -58,15 +61,15 @@ export const ImageWrapperFields: Fields<ImageWrapperProps> = {
   data: YextField(msg("fields.data", "Data"), {
     type: "object",
     objectFields: {
-      image: YextField<any, ImageType | ComplexImageType | AssetImageType>(
-        msg("fields.options.image", "Image"),
-        {
-          type: "entityField",
-          filter: {
-            types: ["type.image"],
-          },
-        }
-      ),
+      image: YextField<
+        any,
+        ImageType | ComplexImageType | TranslatableAssetImage
+      >(msg("fields.options.image", "Image"), {
+        type: "entityField",
+        filter: {
+          types: ["type.image"],
+        },
+      }),
       link: YextField(msg("fields.link", "Link"), {
         type: "translatableString",
       }),
@@ -104,9 +107,17 @@ const ImageWrapperComponent: PuckComponent<ImageWrapperProps> = (props) => {
   }, [parentData, data.image, i18n.language, streamDocument]);
 
   const getImageUrl = (
-    image: ImageType | ComplexImageType | AssetImageType | undefined
+    image: ImageType | ComplexImageType | TranslatableAssetImage | undefined
   ): string | undefined => {
     if (!image) {
+      return undefined;
+    }
+
+    if ("hasLocalizedValue" in image) {
+      const localized = image[i18n.language];
+      if (localized && typeof localized === "object") {
+        return localized.url;
+      }
       return undefined;
     }